test: copy only necessary files for integration test
dnsgit only requires the files found in bin/ and lib/ (plus,
depending on the Ruby installation method, Gemfile and
Gemfile.lock).

The rest is not a runtime dependency and can be ignored for
the integration test.

This circumvents a PEBKAC where a data/ and/or tmp/ directory
exists in the local working directory and which were copied
into the integration test's working directory---which then
lead to very interesting test failures :).

Signed-off-by: Dominik Menke <dom@digineo.de>